feat: Add 'ReactSortableJSByReducer' component.
Add the original 'color' data.
Set the state of filtered data base on the 'checked' property of each 'color' item that is true.
Add a new 'react-sortable-reducer'.
Add the 'onEnd' event in the 'ReactSortable' component.
Dispatch a 'change' action when the 'onEnd' event is triggered to
update the original 'color' data.