feat: Implement infinit scroll using custom hook.
Create a new target variable using the 'useRef' hook to directly reference the DOM element.
Pass the callback function to the 'useIntersectionOserver', and retrive the returned observe and unobserve functions.
Call the 'observe' function when the component mounts in the 'useEffect' hook.
Call the 'unobserve' function to stop observing when component unmounts in the 'useEffect' hook.